Random level generation!
Algo now allows for a small (10%) chance of rooms not appearing. When this happens, the hallways still manage to connect to other rooms WITH HIGH PROBABILITY (not that it's not 100%; there's still a SMALL chance that an important room won't connect, but after testing I can confidently say it's no more than 3%. Will try to improve with future iterations.