Fixed new level generation bug, minor UI adjusts..
Closures for entityMatrix and Map meant matrices could not be refreshed normally. Added this.entityMatrix.clear() and this.map.generate() to work on (scope hidden) matrix objects. Also fixed an error with position for the 'About' screen that left a few pixels of the map and status area visible underneath.